feat:REQ-020326 refactor para el registro de series por sucursal

+++ src/main/java/com/svc/ventas/models/entity/Empresa.java
@@ -16,7 +16,7 @@ public class Empresa implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id_empresa")
-	private Integer idEmpresa;
+	private Long idEmpresa;
 
 	private String ruc;
